develop: modularizing chalice
